Basic Functionality Completion
The backend with methods using flask to upload the audio, list the stories, create tokens for deleting the stories, and liking the stories is completed. There is a database that is created in models.py to store important information for each audio file, which is used especially for the delete funciton. The delete button through token creation in database only works for the story the user created, and no others. The frontend with the stylying and button creation is also completed. There will still be updates that will need to be made like a comment section.